Added Unit Test for formatReceivedTransactionDate

+++ app/src/main/java/com/decagonhq/decapay/feature/logexpense/presentation/LogExpenseBottomSheetFragment.kt
@@ -112,7 +112,7 @@ class LogExpenseBottomSheetFragment : BottomSheetDialogFragment() {
                 logExpenseViewModel.userAddExpense(
                     selectedBudgetId, selectedCategoryId,
                     LogExpenseRequestBody(
-                        amountSpent.toDouble(), description, transactionDate
+                        amountSpent, description, transactionDate
                     )
                 )
             }
